Maintainability... Moved methods around
